pyproject.toml
[build-system]
requires = ["hatchling>=1.24.2"]
build-backend = "hatchling.build"

[project]
name = "infostream"
version = "0.1.0"
description = "Extensible source ingestion and daily digest pipeline"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pydantic>=2.8.0",
    "PyYAML>=6.0.1",
    "httpx>=0.27.0",
    "feedparser>=6.0.11",
    "python-dateutil>=2.9.0.post0",
    "openai>=1.40.0",
    "jinja2>=3.1.4",
    "orjson>=3.10.7",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "respx>=0.21.1",
]

[project.scripts]
infostream = "infostream.cli:main"

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"
